Some bugfixes, one really really weird bug
To reproduce: start a game, find a solution, then leave the game (close tab or reload) before the timer is up.
You may need to uncomment the ajax play loop in line 64 or call ajax("play", {data}, play, solo); in the browser console. 
On re-entering the game (reloading the page), the "solved" variable in the ajax init call response is set to null during JSON.parse.
This affects response.game.json.solved, but not response.test.json.solved, a test copy of the same object. This will not occur if response.game, response.players or response.game.json are renamed. Will require further testing, for now I'll probably just rename the thing.